requirements.txt
groq==0.4.2
pyaudio==0.2.11
numpy==1.24.3
soundfile==0.12.1
websockets==12.0
asyncio-mqtt==0.16.1
python-dotenv==1.0.0 